+++ Welly/RemoteControlContainer.m
@@ -54,7 +54,7 @@
 	return NO;	
 }
 
-- (unsigned int) count {
+- (NSUInteger) count {
 	return [remoteControls count];
 }
