+++ Welly/WLEditingCursorMoveHotspotHandler.m
@@ -63,7 +63,7 @@ static NSCursor *gMoveCursor = nil;
 - (void)mouseUp:(NSEvent *)theEvent {
 	// click to move cursor
 	NSPoint p = [_view convertPoint:[theEvent locationInWindow] fromView:nil];
-	int _selectionLocation = [_view convertIndexFromPoint: p];
+	NSInteger _selectionLocation = [_view convertIndexFromPoint: p];
 	
 	unsigned char cmd[_maxRow * _maxColumn * 3];
 	unsigned int cmdLength = 0;
@@ -72,13 +72,13 @@ static NSCursor *gMoveCursor = nil;
 	// however, since it is enabled by default in smth (switchible by ctrl-p) and disabled in ptt,
 	// we temporarily use bbsType here...
 	if ([ds bbsType] == WLMaple) { // auto-break-line IS NOT enabled in bbs
-		int moveToRow = _selectionLocation / _maxColumn;
-		int moveToCol = _selectionLocation % _maxColumn;
+		NSInteger moveToRow = _selectionLocation / _maxColumn;
+		NSInteger moveToCol = _selectionLocation % _maxColumn;
 		BOOL home = NO;
 		if (moveToRow > [ds cursorRow]) {
 			cmd[cmdLength++] = 0x01;
 			home = YES;
-			for (int i = [ds cursorRow]; i < moveToRow; i++) {
+			for (NSInteger i = [ds cursorRow]; i < moveToRow; i++) {
 				cmd[cmdLength++] = 0x1B;
 				cmd[cmdLength++] = 0x4F;
 				cmd[cmdLength++] = 0x42;
@@ -86,7 +86,7 @@ static NSCursor *gMoveCursor = nil;
 		} else if (moveToRow < [ds cursorRow]) {
 			cmd[cmdLength++] = 0x01;
 			home = YES;
-			for (int i = [ds cursorRow]; i > moveToRow; i--) {
+			for (NSInteger i = [ds cursorRow]; i > moveToRow; i--) {
 				cmd[cmdLength++] = 0x1B;
 				cmd[cmdLength++] = 0x4F;
 				cmd[cmdLength++] = 0x41;
@@ -103,7 +103,7 @@ static NSCursor *gMoveCursor = nil;
 				}
 			}
 		} else if (moveToCol > [ds cursorColumn]) {
-			for (int i = [ds cursorColumn]; i < moveToCol; i++) {
+			for (NSInteger i = [ds cursorColumn]; i < moveToCol; i++) {
 				if (currRow[i].attr.f.doubleByte != 2 || [[[_view frontMostConnection] site] shouldDetectDoubleByte]) {
 					cmd[cmdLength++] = 0x1B;
 					cmd[cmdLength++] = 0x4F;
@@ -111,7 +111,7 @@ static NSCursor *gMoveCursor = nil;
 				}
 			}
 		} else if (moveToCol < [ds cursorColumn]) {
-			for (int i = [ds cursorColumn]; i > moveToCol; i--) {
+			for (NSInteger i = [ds cursorColumn]; i > moveToCol; i--) {
 				if (currRow[i].attr.f.doubleByte != 2 || [[[_view frontMostConnection] site] shouldDetectDoubleByte]) {
 					cmd[cmdLength++] = 0x1B;
 					cmd[cmdLength++] = 0x4F;
@@ -120,12 +120,12 @@ static NSCursor *gMoveCursor = nil;
 			}
 		}
 	} else { // auto-break-line IS enabled in bbs
-		int thisRow = [ds cursorRow];
-		int cursorLocation = thisRow * _maxColumn + [ds cursorColumn];
-		int prevRow = -1;
-		int lastEffectiveChar = -1;
+		NSInteger thisRow = [ds cursorRow];
+		NSInteger cursorLocation = thisRow * _maxColumn + [ds cursorColumn];
+		NSInteger prevRow = -1;
+		NSInteger lastEffectiveChar = -1;
 		if (cursorLocation < _selectionLocation) {
-			for (int i = cursorLocation; i < _selectionLocation; ++i) {
+			for (NSInteger i = cursorLocation; i < _selectionLocation; ++i) {
 				thisRow = i / _maxColumn;
 				if (thisRow != prevRow) {
 					cell *currRow = [ds cellsOfRow:thisRow];
@@ -144,7 +144,7 @@ static NSCursor *gMoveCursor = nil;
 				}
 			}
 		} else {
-			for (int i = cursorLocation; i > _selectionLocation; --i) {
+			for (NSInteger i = cursorLocation; i > _selectionLocation; --i) {
 				thisRow = i / _maxColumn;
 				if (thisRow != prevRow) {
 					cell *currRow = [ds cellsOfRow:thisRow];
